# packages/api-client/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Freshbot Butler API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/auth/household-session": {
      "post": {
        "summary": "Create Household Session",
        "operationId": "createHouseholdSession",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HouseholdSessionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HouseholdSessionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/today": {
      "get": {
        "summary": "Get Today",
        "operationId": "getTodayDashboard",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TodayResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/freshness-overrides": {
      "get": {
        "summary": "List Freshness Overrides",
        "operationId": "listFreshnessOverrides",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FreshnessPolicyListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/freshness-overrides/{category}": {
      "put": {
        "summary": "Save Freshness Override",
        "operationId": "saveFreshnessOverride",
        "parameters": [
          {
            "name": "category",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Category"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FreshnessOverrideRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FreshnessPolicySummary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/text-capture/drafts": {
      "post": {
        "summary": "Create Text Capture Drafts",
        "operationId": "createTextCaptureDrafts",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TextCaptureDraftRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TextCaptureDraftResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/voice-capture/drafts": {
      "post": {
        "summary": "Create Voice Capture Drafts",
        "operationId": "createVoiceCaptureDrafts",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_createVoiceCaptureDrafts"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VoiceCaptureDraftResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/package-photo/drafts": {
      "post": {
        "summary": "Create Package Photo Drafts",
        "operationId": "createPackagePhotoDrafts",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_createPackagePhotoDrafts"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PackagePhotoDraftResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/package-photo/drafts/{capture_id}": {
      "get": {
        "summary": "Get Package Photo Drafts",
        "operationId": "getPackagePhotoDrafts",
        "parameters": [
          {
            "name": "capture_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Capture Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PackagePhotoDraftResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/text-capture/confirm": {
      "post": {
        "summary": "Confirm Text Capture Drafts",
        "operationId": "confirmTextCaptureDrafts",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TextCaptureConfirmRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TextCaptureConfirmResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/package-photo/confirm": {
      "post": {
        "summary": "Confirm Package Photo Drafts",
        "operationId": "confirmPackagePhotoDrafts",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PackagePhotoConfirmRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TextCaptureConfirmResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Body_createPackagePhotoDrafts": {
        "properties": {
          "photo": {
            "type": "string",
            "format": "binary",
            "title": "Photo"
          }
        },
        "type": "object",
        "required": [
          "photo"
        ],
        "title": "Body_createPackagePhotoDrafts"
      },
      "Body_createVoiceCaptureDrafts": {
        "properties": {
          "audio_file": {
            "type": "string",
            "format": "binary",
            "title": "Audio File"
          }
        },
        "type": "object",
        "required": [
          "audio_file"
        ],
        "title": "Body_createVoiceCaptureDrafts"
      },
      "FreshnessOverrideRequest": {
        "properties": {
          "shelf_life_days": {
            "type": "integer",
            "maximum": 365.0,
            "minimum": 0.0,
            "title": "Shelf Life Days"
          },
          "soon_window_days": {
            "type": "integer",
            "maximum": 365.0,
            "minimum": 0.0,
            "title": "Soon Window Days"
          }
        },
        "type": "object",
        "required": [
          "shelf_life_days",
          "soon_window_days"
        ],
        "title": "FreshnessOverrideRequest"
      },
      "FreshnessPolicyListResponse": {
        "properties": {
          "policies": {
            "items": {
              "$ref": "#/components/schemas/FreshnessPolicySummary"
            },
            "type": "array",
            "title": "Policies"
          }
        },
        "type": "object",
        "required": [
          "policies"
        ],
        "title": "FreshnessPolicyListResponse"
      },
      "FreshnessPolicySummary": {
        "properties": {
          "category": {
            "type": "string",
            "title": "Category"
          },
          "default_shelf_life_days": {
            "type": "integer",
            "title": "Default Shelf Life Days"
          },
          "default_soon_window_days": {
            "type": "integer",
            "title": "Default Soon Window Days"
          },
          "shelf_life_days": {
            "type": "integer",
            "title": "Shelf Life Days"
          },
          "soon_window_days": {
            "type": "integer",
            "title": "Soon Window Days"
          },
          "is_override": {
            "type": "boolean",
            "title": "Is Override"
          }
        },
        "type": "object",
        "required": [
          "category",
          "default_shelf_life_days",
          "default_soon_window_days",
          "shelf_life_days",
          "soon_window_days",
          "is_override"
        ],
        "title": "FreshnessPolicySummary"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "HouseholdSessionRequest": {
        "properties": {
          "household_name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Household Name"
          },
          "member_name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Member Name"
          }
        },
        "type": "object",
        "required": [
          "household_name",
          "member_name"
        ],
        "title": "HouseholdSessionRequest"
      },
      "HouseholdSessionResponse": {
        "properties": {
          "token": {
            "type": "string",
            "title": "Token"
          },
          "locale": {
            "type": "string",
            "title": "Locale"
          },
          "household": {
            "$ref": "#/components/schemas/HouseholdInfo"
          },
          "member": {
            "$ref": "#/components/schemas/MemberInfo"
          }
        },
        "type": "object",
        "required": [
          "token",
          "locale",
          "household",
          "member"
        ],
        "title": "HouseholdSessionResponse"
      },
...
      "TextCaptureConfirmRequest": {
        "properties": {
          "drafts": {
            "items": {
              "$ref": "#/components/schemas/TextCaptureDraft"
            },
            "type": "array",
            "minItems": 1,
            "title": "Drafts"
          }
        },
        "type": "object",
        "required": [
          "drafts"
        ],
        "title": "TextCaptureConfirmRequest"
      },
      "TextCaptureConfirmResponse": {
        "properties": {
          "batches": {
            "items": {
              "$ref": "#/components/schemas/BatchSummary"
            },
            "type": "array",
            "title": "Batches"
          }
        },
        "type": "object",
        "required": [
          "batches"
        ],
        "title": "TextCaptureConfirmResponse"
      },
...
      "TextCaptureDraftRequest": {
        "properties": {
          "input_text": {
            "type": "string",
            "maxLength": 1000,
            "minLength": 1,
            "title": "Input Text"
          }
        },
        "type": "object",
        "required": [
          "input_text"
        ],
        "title": "TextCaptureDraftRequest"
      },
      "TextCaptureDraftResponse": {
        "properties": {
          "drafts": {
            "items": {
              "$ref": "#/components/schemas/TextCaptureDraft"
            },
            "type": "array",
            "title": "Drafts"
          },
          "available_categories": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Available Categories"
          },
          "available_locations": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Available Locations"
          }
        },
        "type": "object",
        "required": [
          "drafts",
          "available_categories",
          "available_locations"
        ],
        "title": "TextCaptureDraftResponse"
      },
      "TodayResponse": {
        "properties": {
          "household_name": {
            "type": "string",
            "title": "Household Name"
          },
          "member_name": {
            "type": "string",
            "title": "Member Name"
          },
          "locale": {
            "type": "string",
            "title": "Locale"
          },
          "sections": {
            "$ref": "#/components/schemas/TodaySections"
          }
        },
        "type": "object",
        "required": [
          "household_name",
          "member_name",
          "locale",
          "sections"
        ],
        "title": "TodayResponse"
      },
...
      "VoiceCaptureDraftResponse": {
        "properties": {
          "drafts": {
            "items": {
              "$ref": "#/components/schemas/TextCaptureDraft"
            },
            "type": "array",
            "title": "Drafts"
          },
          "available_categories": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Available Categories"
          },
          "available_locations": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Available Locations"
          },
          "transcript": {
            "type": "string",
            "title": "Transcript"
          }
        },
        "type": "object",
        "required": [
          "drafts",
          "available_categories",
          "available_locations",
          "transcript"
        ],
        "title": "VoiceCaptureDraftResponse"
      }
    }
  }
}