# packages/api-client/src/generated/models/Body_createVoiceCaptureDrafts.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type Body_createVoiceCaptureDrafts = {
    audio_file: Blob;
};